+++ src/components/Toolbar/Toolbar.story.js
@@ -5,39 +5,144 @@ import { withKnobs } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 
 import Toolbar from './index.js';
+import Dropdown from '../Dropdown/index.js';
+import Button from '../Button/index.js';
+
+const envData = [
+  { title: 'Development' },
+  { title: 'Staging' },
+  { title: 'Production' },
+];
+
+const statusData = [
+  { title: 'Running' },
+  { title: 'Active' },
+  { title: 'Paused' },
+];
 
 const stories = storiesOf('Toolbar', module);
-stories
-  .addDecorator(withKnobs)
-  .addDecorator(story => (
-    <div id="root-preview">
-      {story()}
-    </div>
-  ));
+stories.addDecorator(withKnobs).addDecorator((story) => (
+  <div id="root-preview" style={{ height: '300px' }}>
+    {story()}
+  </div>
+));
 
 stories
-  .add('Default', (() => (<div>
-    <Toolbar>
-      <Toolbar.Left>
-        <Toolbar.Link
-          href={ 'www.google.com' }
-          label='Link'
-        />
-      </Toolbar.Left>
-      <Toolbar.Right>
-        <Toolbar.Link
-          isDisabled={ true }
-          href={ 'www.google.com' }
-          label='Disabled Link'
-        />
-        <Toolbar.Link
-          href={ 'www.google.com' }
-          label='Enabled Link'
-        />
-        <Toolbar.Button
-          onClick={ action('Saying hi!') }
-          label='Button'
-        />
-      </Toolbar.Right>
-    </Toolbar>
-  </div>)));
+  .add('Default', () => (
+    <div>
+      <Toolbar>
+        <Toolbar.Left>
+          <Toolbar.Link href={ 'www.google.com' } label="Link" />
+        </Toolbar.Left>
+        <Toolbar.Right>
+          <Toolbar.Link
+            isDisabled={ true }
+            href={ 'www.google.com' }
+            label="Disabled Link"
+          />
+          <Toolbar.Link href={ 'www.google.com' } label="Enabled Link" />
+          <Toolbar.Button onClick={ action('Saying hi!') } label="Button" />
+        </Toolbar.Right>
+      </Toolbar>
+    </div>
+  ))
+  .add('Without bottom border and shadow', () => (
+    <div>
+      <Toolbar toolbarStyle="bare">
+        <Toolbar.Left>
+          <Toolbar.Link href={ 'www.google.com' } label="Link" />
+        </Toolbar.Left>
+        <Toolbar.Right>
+          <Toolbar.Link
+            isDisabled={ true }
+            href={ 'www.google.com' }
+            label="Disabled Link"
+          />
+          <Toolbar.Link href={ 'www.google.com' } label="Enabled Link" />
+          <Toolbar.Button onClick={ action('Saying hi!') } label="Button" />
+        </Toolbar.Right>
+      </Toolbar>
+    </div>
+  ))
+  .add('With Dropdowns and styled Buttons', () => (
+    <div>
+      <Toolbar toolbarStyle="bare">
+        <Toolbar.Left>
+          <Dropdown
+            buttonContent={{
+              label: 'Environment',
+              content: 'Production',
+            }}
+            style="plain"
+            arrowIcon="down">
+            <Dropdown.Contents>
+              {envData.map((item, index) => {
+                return (
+                  <Dropdown.ListItem key={ index } role="option">
+                    <Dropdown.BlockLink
+                      value={ item.title }
+                      onClick={ action('click on complex item') }>
+                      <Dropdown.BlockLinkText text={ item.title } />
+                      <Dropdown.BlockLinkSecondaryText
+                        secondaryText={ item.description }
+                      />
+                    </Dropdown.BlockLink>
+                  </Dropdown.ListItem>
+                );
+              })}
+            </Dropdown.Contents>
+          </Dropdown>
+          <Dropdown
+            buttonContent={{
+              label: 'Status',
+              content: 'Active',
+            }}
+            style="plain"
+            arrowIcon="down">
+            <Dropdown.Contents>
+              {envData.map((item, index) => {
+                return (
+                  <Dropdown.ListItem key={ index } role="option">
+                    <Dropdown.BlockLink
+                      value={ item.title }
+                      onClick={ action('click on complex item') }>
+                      <Dropdown.BlockLinkText text={ item.title } />
+                      <Dropdown.BlockLinkSecondaryText
+                        secondaryText={ item.description }
+                      />
+                    </Dropdown.BlockLink>
+                  </Dropdown.ListItem>
+                );
+              })}
+            </Dropdown.Contents>
+          </Dropdown>
+          <Dropdown
+            buttonContent={{
+              label: 'Rule Type',
+              content: 'Any',
+            }}
+            style="plain"
+            arrowIcon="down">
+            <Dropdown.Contents>
+              {statusData.map((item, index) => {
+                return (
+                  <Dropdown.ListItem key={ index } role="option">
+                    <Dropdown.BlockLink
+                      value={ item.title }
+                      onClick={ action('click on complex item') }>
+                      <Dropdown.BlockLinkText text={ item.title } />
+                    </Dropdown.BlockLink>
+                  </Dropdown.ListItem>
+                );
+              })}
+            </Dropdown.Contents>
+          </Dropdown>
+        </Toolbar.Left>
+        <Toolbar.Right>
+          <Button style="highlight" onClick={ action('Saying hi!') }>
+            Create New
+          </Button>
+        </Toolbar.Right>
+      </Toolbar>
+    </div>
+  ));