+++ src/components/Toolbar/index.js
@@ -2,36 +2,8 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import classNames from 'classnames';
 
-const Toolbar = (props) => {
-  const toolbarContentClasses = classNames({
-    toolbar__content: true,
-    'background--white': props.isBottomToolbar,
-    'border--top': props.isBottomToolbar,
-    'no-border--bottom': props.isBottomToolbar,
-    'hard--left': props.isBottomToolbar,
-  });
-
-  return (
-    <div
-      data-oui-component={ true }
-      className="toolbar"
-      data-test-section={ props.testSection }>
-      <div className={ toolbarContentClasses }>
-        { props.children }
-      </div>
-    </div>
-  );
-};
-
-Toolbar.defaultProps = {
-  isBottomToolbar: false,
-};
-
-Toolbar.propTypes = {
-  children: PropTypes.node.isRequired,
-  isBottomToolbar: PropTypes.bool,
-  testSection: PropTypes.string,
-};
+import ToolbarButton from './ToolbarButton';
+import ToolbarLink from './ToolbarLink';
 
 const Left = props => {
   return (
@@ -53,102 +25,41 @@ Right.propTypes = {
   children: PropTypes.node.isRequired,
 };
 
-const ToolbarButtonContents = props => (
-  <div className="flex flex-align--center editor__select-size__text">
-    { props.icon && (
-      <svg className="lego-icon push--right vertical-align--middle">
-        <use xlinkHref={ props.icon } />
-      </svg>
-    )}
-    <div>
-      { props.title && (
-        <div className="editor__select-size__value">
-          { props.title }
-        </div>
-      )}
+const Toolbar = (props) => {
+  const toolbarContentClasses = classNames({
+    'toolbar__content': true,
+    'toolbar__content--bare': props.toolbarStyle === 'bare',
+    'background--white': props.isBottomToolbar,
+    'border--top': props.isBottomToolbar,
+    'no-border--bottom': props.isBottomToolbar || props.toolbarStyle === 'bare',
+    'hard--left': props.isBottomToolbar,
+  });
 
-      <span>{ props.label }</span>
+  return (
+    <div
+      data-oui-component={ true }
+      className="toolbar"
+      data-test-section={ props.testSection }>
+      <div className={ toolbarContentClasses }>
+        { props.children }
+      </div>
     </div>
-
-    { props.isDropdown && (
-      <span className="lego-arrow-inline--down push--left" />
-    )}
-  </div>
-);
-
-ToolbarButtonContents.propTypes = {
-  icon: PropTypes.string,
-  isDropdown: PropTypes.bool,
-  label: PropTypes.oneOfType([
-    PropTypes.string.isRequired,
-    PropTypes.node.isRequired,
-  ]).isRequired,
-  title: PropTypes.string,
+  );
 };
 
-const ToolbarLink = (props, context) => (
-  <a
-    className={ classNames({
-      flex: true,
-      toolbar__button: true,
-      'is-active': props.isActive,
-      ['link--disabled']: props.isDisabled,
-      'pointer-events--none': props.isDisabled,
-    }) }
-    data-test-section={ props.testSection }
-    href={ props.href }>
-    { ToolbarButtonContents(props) }
-  </a>
-);
+Toolbar.defaultProps = {
+  isBottomToolbar: false,
+};
 
-ToolbarLink.propTypes = {
-  href: PropTypes.string,
-  icon: PropTypes.string,
-  isActive: PropTypes.bool,
-  isDisabled: PropTypes.bool,
-  isDropdown: PropTypes.bool,
-  label: PropTypes.oneOfType([
-    PropTypes.string.isRequired,
-    PropTypes.node.isRequired,
-  ]).isRequired,
-  onClick: PropTypes.func,
+Toolbar.propTypes = {
+  /** Items to render for the toolbar */
+  children: PropTypes.node.isRequired,
+  /** Whether this toolbar sits at the bottom of a page */
+  isBottomToolbar: PropTypes.bool,
+  /** Hook to uze for automated testing */
   testSection: PropTypes.string,
-  title: PropTypes.string,
-};
-
-const ToolbarButton = (props, context) => (
-  <button
-    className={ classNames({
-      toolbar__button: true,
-      'is-active': props.isActive,
-    }) }
-    type="button"
-    disabled={ props.isDisabled }
-    data-test-section={ props.testSection }
-    data-track-id={ props.testSection }
-    onClick={ props.onClick }>
-    { ToolbarButtonContents(props) }
-  </button>
-);
-
-ToolbarButton.propTypes = {
-  icon: PropTypes.string,
-  isActive: PropTypes.bool,
-  isDisabled: PropTypes.bool,
-  isDropdown: PropTypes.bool,
-  label: PropTypes.oneOfType([
-    PropTypes.node.isRequired,
-    PropTypes.string.isRequired,
-  ]).isRequired,
-  onClick: PropTypes.func,
-  testSection: PropTypes.string,
-  title: PropTypes.string,
-};
-
-ToolbarButton.defaultProps = {
-  isDropdown: false,
-  isActive: false,
-  isDisabled: false,
+  /** The style to use for this toolbar */
+  toolbarStyle: PropTypes.oneOf(['bare']),
 };
 
 Toolbar.Button = ToolbarButton;